SEBA/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
pdfplumber==0.11.4
pandas==2.2.2
supabase==2.7.4
python-dotenv==1.0.1
httpx==0.27.2
